+++ CHANGELOG.md
@@ -6,6 +6,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/).
 ### Changed
 - `ImmutableCollection`, `ImmutableList`, `ImmutableMap`, `ImmutableSet`: added clarifying JavaDoc
 - `MiniMap0`, `MiniSet0`: minor optimisations
+- `pom.xml`: fixed repository reference
 
 ## [2.3] - 2023-01-08
 ### Changed